filetype: Set diff filetype for COMMIT_EDITMSG files
This filename is set when writing a commit message in git.

If you run git-commit with an --verbose option or if you just have set
it in your git-config, that shows you the patch diff on bottom
of the file.

Unfortunately, now every patch file will set colorcolumn, but overall
I think this is better than having duplicated lexers.