lua: filetype: use alt_name field to alias to other lexers
This obviously draws on the alt_name parameter used in scintillua
but rather then passing it all the way up to the chain to
lexers.load() we will just handle it in set_filetype().

8a420ec accidently readded detection for the removed git-commit
lexer which somehow went unnoticed until we readded caching for
the new lexers. Instead of just removing it we can alias to the
diff filetype and only set the colorcolumn for commit messages.
This fixes the incorrect behaviour of adding a colorcolumn to diff
and patch files and thus completely reverts 0cc684f.