destroy the correct subprocess
When a new subprocess is created during an EXIT event of
another subprocess new_process_in_pool will update the
process_pool pointer. Since we use a pointer to a pointer
for iterating all processes during vis_process_tick its
value will be different before executing the event and after
creating the new subprocess. This causes the updated pointer
to be erroneously destroyed and leaves the Process of the
reaped child behind which causes consecutive waitpid calls
to fail with ECHILD.

This is fixed by destroying the proper current subprocess
and updating the iteration pointer accordingly.

Fixes: 0b015320382e74fcb385a46a81304f588ed27f77