remove some ui pointer chasing
There only exists a single Ui so there is no need to force a
pointer redirection for accessing it.

The Ui member was moved down in vis-core.h to punt around an issue
with the way lua checks for existing objects. It may show up again
as I flatten more structs.