support piping a buffer to an external process
Currently only Text objects can be piped to external commands.
This is tedious if data not available in any file should be passed
to an external process (e.g. building options and passing them to
vis-menu).

This adds the option to pass a buffer to _vis_pipe and provides wrapper
functions for the original behavior and the new one.