remove the vis->initialized member
I already fixed the reason that this even existed (vis_event_emit
getting called at random times when the editor wasn't ready).

The option checking in main() was moved up because I noticed it
was in the wrong place while thinking about where to emit the INIT
event. There is no reason to do a bunch of useless work just to
print the version.