fix support for incremental rebuilds as set out by 436044
the POSIX Makefile macro ".c.o"(reintroduced into this patch), for
example, compiles text.c -> text.o, if text.o is either older, or it
doesn't exist.

this was removed in patches 4ca7119 & 09ba77a, after which, if you
modify any C file, vis would not be re-compiled.

the goal of 4ca7119 was to stop the base directory from being polluted
with .o & .d files. This patch polluates the base directory.

the only simple alternative to .c.o, which doesn't pollute, is the
following:

obj/%.o: %.c
	${CC} ${CFLAGS} ${CFLAGS_VIS} -c $< -o $@

this is not POSIX, however.