Increase the default LPeg stack size
It is not clear how large this value should be. It appears to be a function of grammar complexity.

For example, a problematic HTML file requires a value of 1329 to work, but removing either the
'attribute' rule or an embedded lexer reduces the limit.

However, identifying which files trigger a stack overflow is not trivial. It does not appear to
depend on file size. For example, the problematic HTML file is 125K, but a non-problematic HTML
file of 500K works.