util: replace memrchr with internal version
The amount of code we need to detect if this is present and handle
the fallback is more than if we just provide it ourselves.

We are passing in a difference of pointers so the argument type
should be ptrdiff_t. This avoids a C brain damage of having
unsigned size type.