vis-lua: add file:snapshot method
A number of plugins, for example vis-spellcheck[0], want to
perform bulk edits on the file text. They may desire for those
edits to be split into multiple undo/redo points. This is
achievable by switching the mode between insert and normal mode
but that has other side effects. Instead introduce a method of
doing this directly.

[0]: https://gitlab.com/muhq/vis-spellcheck