text: remove idiotic dirname/basename usage in atomic save path
these are two of the worst functions defined by POSix. We do not
want to be doing extra memory allocations in the atomic save path
just to appease their terrible design. This change removes two of
the extra allocations in this path with final one being possible
to remove with a little rework.